Separate FilterInput variants into their own TUs
