+++ entity/BUILD.gn
@@ -39,14 +39,20 @@ impeller_component("entity") {
     "contents/contents.h",
     "contents/filters/blend_filter_contents.cc",
     "contents/filters/blend_filter_contents.h",
-    "contents/filters/filter_contents.cc",
-    "contents/filters/filter_contents.h",
-    "contents/filters/filter_input.cc",
-    "contents/filters/filter_input.h",
-    "contents/filters/gaussian_blur_filter_contents.cc",
-    "contents/filters/gaussian_blur_filter_contents.h",
     "contents/filters/border_mask_blur_filter_contents.cc",
     "contents/filters/border_mask_blur_filter_contents.h",
+    "contents/filters/filter_contents.cc",
+    "contents/filters/filter_contents.h",
+    "contents/filters/gaussian_blur_filter_contents.cc",
+    "contents/filters/gaussian_blur_filter_contents.h",
+    "contents/filters/inputs/contents_filter_input.cc",
+    "contents/filters/inputs/contents_filter_input.h",
+    "contents/filters/inputs/filter_contents_filter_input.cc",
+    "contents/filters/inputs/filter_contents_filter_input.h",
+    "contents/filters/inputs/filter_input.cc",
+    "contents/filters/inputs/filter_input.h",
+    "contents/filters/inputs/texture_filter_input.cc",
+    "contents/filters/inputs/texture_filter_input.h",
     "contents/linear_gradient_contents.cc",
     "contents/linear_gradient_contents.h",
     "contents/snapshot.cc",
