+++ entity/contents/filters/gaussian_blur_filter_contents.h
@@ -7,7 +7,7 @@
 #include <memory>
 #include <optional>
 #include "impeller/entity/contents/filters/filter_contents.h"
-#include "impeller/entity/contents/filters/filter_input.h"
+#include "impeller/entity/contents/filters/inputs/filter_input.h"
 
 namespace impeller {
